==> notebooks/01 Exploration.ipynb <==
import pandas as pd
import numpy as np
pd.options.display.max_columns = None

df = pd.read_csv('../raw/loan_data_2007_2014.csv')
df.head()

df.info()

# Preprocessing

df_preprocessed = df.copy()

**emp_length_int**

df['emp_length_int'] = df.emp_length.str.replace('< 1 year','0')
df['emp_length_int'] = df.emp_length_int.fillna('0')
df['emp_length_int'] = df.emp_length_int.str.replace('+','',regex=False).str.replace('years','').str.replace('year','').str.replace(' ','')
df.emp_length_int = pd.to_numeric(df.emp_length_int)
df.emp_length_int.describe()

df.earliest_cr_line.head()

**earliest_cr_line**

df['earliest_cr_line_proc'] = pd.to_datetime(df.earliest_cr_line,format='%b-%y')
df['earliest_cr_line_months'] = (pd.to_datetime('01-01-2018') - df.earliest_cr_line_proc)/np.timedelta64(1,'M')
# df.earliest_cr_line.head()
df.earliest_cr_line_months.describe()

df.loc[df.earliest_cr_line_months<0,['earliest_cr_line','earliest_cr_line_months','earliest_cr_line_proc']]

df.loc[df.earliest_cr_line_months<0,'earliest_cr_line_months'] = df.loc[:,'earliest_cr_line_months'].max()

df.earliest_cr_line_months.describe()

**term**

df['term_int'] = df.term.str.replace(' ','').str.replace('months','').astype(int)
df.term_int.unique()

**issue_d**

df['issue_d_date'] = pd.to_datetime(df['issue_d'],format='%b-%y')
df['mths_since_issue_d'] = (pd.to_datetime('01-12-2017') - df.issue_d_date)/np.timedelta64(1,'M')
df['mths_since_issue_d'] = df.mths_since_issue_d.astype(int)
df.mths_since_issue_d.describe()

**discrete**

loan_data_dummies = [pd.get_dummies(df['grade'], prefix = 'grade', prefix_sep = '_'),
                     pd.get_dummies(df['sub_grade'], prefix = 'sub_grade', prefix_sep = '_'),
                     pd.get_dummies(df['home_ownership'], prefix = 'home_ownership', prefix_sep = '_'),
                     pd.get_dummies(df['verification_status'], prefix = 'verification_status', prefix_sep = '_'),
                     pd.get_dummies(df['loan_status'], prefix = 'loan_status', prefix_sep = '_'),
                     pd.get_dummies(df['purpose'], prefix = 'purpose', prefix_sep = '_'),
                     pd.get_dummies(df['addr_state'], prefix = 'addr_state', prefix_sep = '_'),
                     pd.get_dummies(df['initial_list_status'], prefix = 'initial_list_status', prefix_sep = '_')]
dummy_df = pd.concat(loan_data_dummies,axis=1)
df = pd.concat([df,dummy_df],axis=1)

**fillna**

df.total_rev_hi_lim = df.total_rev_hi_lim.fillna(df.funded_amnt) 

df.annual_inc.fillna(df.annual_inc.mean(),inplace=True)
vars_ = ['earliest_cr_line_months','acc_now_delinq','total_acc',
         'pub_rec','open_acc','inq_last_6mths','delinq_2yrs','emp_length_int']
for var in vars_:
    df[var].fillna(0,inplace=True)

df.loan_status.value_counts()

**good_bad**

df['good_bad'] = np.where(df['loan_status'].isin(['Charged Off','Default','Does not meet the credit policy. Status: Charged Off',
                                                  'Late (16-30 days)']),0,1)

df.to_csv('../processed/01_data_preprocessed.csv',index=False)
