==> notebooks/02 Train test split.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.model_selection import train_test_split

pd.options.display.max_columns = None

sns.set()

### Data

df = pd.read_csv('../processed/01_data_preprocessed.csv')
df.head()

df_train, df_test, y_train, y_test = train_test_split(df.drop('good_bad',axis=1),
                                                      df.good_bad, test_size = 0.2, random_state=42)

df_train.to_csv('../processed/x_train.csv',index=False)
df_test.to_csv('../processed/x_test.csv',index=False)
y_train.to_csv('../processed/y_train.csv',index=False)
y_test.to_csv('../processed/y_test.csv',index=False)
