==> notebooks/07 Scorecard.ipynb <==
import pandas as pd
import numpy as np

from sklearn.metrics import roc_curve

pd.options.display.max_columns = None
pd.options.display.max_rows = 100

# Applying the PD Model

summary_table = pd.read_csv('../processed/05_summary_table.csv')
summary_table.head()

ref_categories = ['grade:G',
'home_ownership:RENT_OTHER_NONE_ANY',
'addr_state:ND_NE_IA_NV_FL_HI_AL',
'verification_status:Verified',
'purpose:educ__sm_b__wedd__ren_en__mov__house',
'initial_list_status:f',
'term:60',
'emp_length:0',
'mths_since_issue_d:>84',
'int_rate:>20.281',
'mths_since_earliest_cr_line:<140',
'inq_last_6mths:>6',
'acc_now_delinq:0',
'annual_inc:<20K',
'dti:>35',
'mths_since_last_delinq:0-3',
'mths_since_last_record:0-2']

### Creating a Scorecard

df_ref_categories = pd.DataFrame(ref_categories, columns = ['Feature name'])
df_ref_categories['Coefficients'] = 0
df_ref_categories['p_values'] = np.nan
df_ref_categories

df_scorecard = pd.concat([summary_table, df_ref_categories])
df_scorecard = df_scorecard.sort_values('Feature name').reset_index()

def feature_name(row):
    if(row.startswith('acc_now_delinq')): return 'acc_now_delinq'
    if(row.startswith('grade')): return 'grade'
    if(row.startswith('addr_state')): return 'addr_state'
    if(row.startswith('annual_inc')): return 'annual_inc'
    if(row.startswith('dti')): return 'dti'
    if(row.startswith('emp_length')): return 'emp_length'
    if(row.startswith('home_ownership')): return 'home_ownership'
    if(row.startswith('initial_list_status')): return 'initial_list_status'
    if(row.startswith('inq_last')): return 'inq_last'
    if(row.startswith('int_rate')): return 'int_rate'
    if(row.startswith('mths_since_earliest_cr_line')): return 'mths_since_earliest_cr_line'
    if(row.startswith('mths_since_last_record')): return 'mths_since_last_record'
    if(row.startswith('purpose')): return 'purpose'
    if(row.startswith('term')): return 'term'
    if(row.startswith('verification_status')): return 'verification_status'
    if(row.startswith('mths_since_issue_d')): return 'mths_since_issue_d'
    if(row.startswith('mths_since_last_delinq')): return 'mths_since_last_delinq'
    return row

df_scorecard['Original feature name'] = df_scorecard['Feature name'].apply(feature_name)
df_scorecard['Original feature name'].unique()

min_score = 300
max_score = 850

df_scorecard.groupby('Original feature name')['Coefficients'].min()

min_sum_coef = df_scorecard.groupby('Original feature name')['Coefficients'].min().sum()
min_sum_coef

max_sum_coef = df_scorecard.groupby('Original feature name')['Coefficients'].max().sum()
max_sum_coef

df_scorecard['Score - Calculation'] = df_scorecard['Coefficients'] * (max_score - min_score) / (max_sum_coef - min_sum_coef)

df_scorecard.loc[0,'Score - Calculation'] = ((df_scorecard['Coefficients'][0] - min_sum_coef) / (max_sum_coef - min_sum_coef)) * (max_score - min_score) + min_score

df_scorecard['Score - Preliminary'] = df_scorecard['Score - Calculation'].round()

min_sum_score_prel = df_scorecard.groupby('Original feature name')['Score - Preliminary'].min().sum()
min_sum_score_prel

max_sum_score_prel = df_scorecard.groupby('Original feature name')['Score - Preliminary'].max().sum()
max_sum_score_prel

df_scorecard['Score - Final'] = df_scorecard['Score - Preliminary']
df_scorecard.loc[0,'Score - Final'] = 451

df_scorecard.loc[46,'Score - Final'] = 65

df_scorecard.groupby('Original feature name')['Score - Final'].min().sum(), df_scorecard.groupby('Original feature name')['Score - Final'].max().sum()

# Score prediction

df_scorecard.head(1)

x_test = pd.read_csv('../processed/x_test_04_output.csv')
x_test.insert(0, 'Intercept', 1)

x_test.head(1)

inputs_test_with_ref_cat_w_intercept = x_test[df_scorecard[df_scorecard['Score - Final']!=0]['Feature name'].values]

scorecard_scores = df_scorecard[df_scorecard['Score - Final']!=0][['Score - Final']]

y_scores = inputs_test_with_ref_cat_w_intercept.dot(scorecard_scores.values)

# Cutpoints

sum_coef_from_score = ((y_scores - min_score) / (max_score - min_score)) * (max_sum_coef - min_sum_coef) + min_sum_coef

y_hat_proba_from_score = np.exp(sum_coef_from_score) / (np.exp(sum_coef_from_score) + 1)
y_hat_proba_from_score.head(5)

y_test = pd.read_csv('../processed/06_predictions.csv')
y_test = y_test.sort_values('index')
y_test.head()

fpr, tpr, thresholds = roc_curve(y_test.y_test,y_test.pred_2)
df_cutoffs = pd.concat([pd.DataFrame(thresholds), pd.DataFrame(fpr), pd.DataFrame(tpr)], axis = 1)
df_cutoffs.columns = ['thresholds', 'fpr', 'tpr']
df_cutoffs.loc[0,'thresholds'] = 1 - 1/np.power(10,16)

df_cutoffs['Score'] = ((np.log(df_cutoffs['thresholds'] / (1 - df_cutoffs['thresholds'])) - min_sum_coef) * ((max_score - min_score) / (max_sum_coef - min_sum_coef)) + min_score).round()
df_cutoffs.loc[0,'Score'] = max_score

df_cutoffs.head()
df_cutoffs.tail()

def n_approved(p):
    return np.where(y_test['pred_2'] >= p, 1, 0).sum()

df_cutoffs['N Approved'] = df_cutoffs['thresholds'].apply(n_approved)
df_cutoffs['N Rejected'] = y_test['pred_2'].shape[0] - df_cutoffs['N Approved']
df_cutoffs['Approval Rate'] = df_cutoffs['N Approved'] / y_test['pred_2'].shape[0]
df_cutoffs['Rejection Rate'] = 1 - df_cutoffs['Approval Rate']

df_cutoffs.loc[2000:2020]

df_cutoffs.to_csv('../processed/07_cutoffs.csv',index=False)
df_scorecard.to_csv('../processed/07_scorecard.csv',index=False)
